Implement a new, callback-based, lowering for calls with compounds.
The difference between ABIs is so large that we cannot distingush between
them with flags. Basically, we would have to implement all ABIs in
lower_calls.c. Instead, we let the backends specify how parameter and
result types are to be lowered.

For now, all backends use a dummy ABI, which is the old Firm default,
i.e. the SPARC ABI.